+++ src/components/Project/Project.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import mapStoreToProps from '../../redux/mapStoreToProps';
 //imported Components
-import ProjectTitle from './ProjectTitle';
+import ProjectTitle from './ProjectTitle/ProjectTitle';
