+++ src/components/Project/ProjectTitle/ProjectTitle.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import mapStoreToProps from '../../redux/mapStoreToProps';
-import TaskList from './TaskList';
+import mapStoreToProps from '../../../redux/mapStoreToProps';
+import TaskList from '../TaskList';
+import './ProjectTitle.css'
 
 
 const getCookie = (cookieName) => {
